fixed filename extensions and bug where not updating
