+++ src/app/store/Settings.js
